tenant_object_config/requirements.txt
PyYAML==24.0